Extract the event factory.
